[NO-JIRA] Refactor build process
